+++ blocks/breadcrumb/breadcrumb.css
@@ -1,0 +1,60 @@
+.breadcrumb > div > div > div {
+  display: flex;
+  align-items: center;
+  margin-bottom: 1rem;
+}
+
+.breadcrumb .home {
+  margin-right: 10px;
+  display: inline-flex;
+}
+
+.breadcrumb > div > div > div > span {
+  display: inline-flex;
+  width: 20px;
+}
+
+.breadcrumb > div > div > div > span svg {
+  fill: var(--highlight-color);
+}
+
+.breadcrumb .icon-chevron {
+  transform: rotate(270deg);
+}
+
+.breadcrumb .category-link-btn {
+  --bg-color: inherit;
+  --border-color: inherit;
+  --text-color: inherit;
+
+  background: var(--bg-color);
+  border: 1px solid var(--border-color);
+  color: var(--text-color);
+  border-radius: 50px;
+  font-size: var(--body-font-size-xxs);
+  text-align: center;
+  line-height: 1.1;
+  padding: 0.3rem 0.6rem;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
+
+.breadcrumb .category-link-btn:not(:last-child) {
+  background-color: var(--text-color-inverted);
+}
+
+.breadcrumb .category-link-btn:last-child {
+  color: var(--text-color-inverted);
+}
+
+@media (min-width: 768px) {
+  .breadcrumb .category-link-btn {
+    padding: 0.3rem 0.7rem;
+  }
+}
+
+@media (min-width: 1200px) {
+  .breadcrumb .category-link-btn {
+    font-size: var(--body-font-size-xxxs);
+  }
+}
