+++ scripts/lib-franklin.js
@@ -607,71 +607,6 @@ export function setup() {
 }
 
 /**
- * Convert snake case to title case
- * @param str
- * @returns {*}
- */
-function convertToTitleCase(str) {
-  const words = str.split('-');
-  const capitalizedWords = words.map((word) => word.charAt(0).toUpperCase() + word.slice(1));
-  return capitalizedWords.join(' ');
-}
-
-/**
- * @typedef CrumbData
- * @property {string} url Full URL to which clicking the crumb will redirect.
- * @property {string} path Name of the crumb as it will appear on its label.
- * @property {string} color Name of the color in which the breadcrumb should
- *  be rendered.
- */
-
-/**
- * Adds the given breadcrumb data to an HTML element, styling it for
- * use in the site.
- * @param {Element} target HTML element to which the breadcrumbs will be
- *  added.
- * @param {Array<CrumbData>} crumbData Information about the crumbs to add.
- * @returns {Promise} Resolves when the crumbs have been added.
- */
-export async function buildBreadCrumbs(target, crumbData) {
-  // Use the last item in the list's color
-  const { color } = crumbData[crumbData.length - 1];
-
-  const homeLink = document.createElement('a');
-  homeLink.classList.add('home');
-  homeLink.href = '/';
-  homeLink.innerHTML = '<span class="icon icon-home"></span>';
-  target.append(homeLink);
-
-  crumbData.forEach((crumb, i) => {
-    if (i > 0) {
-      const chevron = document.createElement('span');
-      chevron.innerHTML = '<span class="icon icon-chevron"></span>';
-      target.append(chevron);
-    }
-    const linkButton = document.createElement('a');
-    linkButton.href = crumb.url;
-    linkButton.innerText = convertToTitleCase(crumb.path);
-    linkButton.classList.add('category-link-btn');
-    if (i === crumbData.length - 1) {
-      // linkButton.classList.add(`${color}`);
-      linkButton.style.setProperty('--bg-color', `var(--color-${color})`);
-      linkButton.style.setProperty('--border-color', 'inherit');
-      linkButton.style.setProperty('--text-color', 'inherit');
-    } else {
-      // linkButton.classList.add(`${color}-border`, `${color}-color`);
-      linkButton.style.setProperty('--bg-color', 'inherit');
-      linkButton.style.setProperty('--border-color', `var(--color-${color})`);
-      linkButton.style.setProperty('--text-color', `var(--color-${color})`);
-    }
-
-    target.append(linkButton);
-  });
-
-  return decorateIcons(target);
-}
-
-/**
  * Auto initializiation.
  */
 function init() {
